style: #6 게시글 작성 화면 스타일 분리
