+++ src/components/post/HashtagInputbox.tsx
@@ -1,27 +1,18 @@
-import React from 'react';
-import {TextInput, StyleSheet, Platform} from 'react-native';
+import React, {Component} from 'react';
+import {TextInput} from 'react-native';
+import styles from '../../styles/post/HashtagInputbox';
 
-const InputboxID = () => {
-    const styles = StyleSheet.create({
-        inputStyle: {
-            marginTop: Platform.OS == 'ios' ? 5 : 0,
-            marginLeft: 20,
-            marginBottom: Platform.OS == 'ios' ? 5 : 0,
-            padding: 8,
-            backgroundColor: 'white',
-            borderColor: '#FFFFF',
-            borderRadius: 8
-        }
-    });
-
-    return (
-        <TextInput
-            style = {styles.inputStyle}
-            placeholder = '해쉬태그'
-            placeholderTextColor = {'#909090'}
-        >
-        </TextInput>
-    );
+class HashtagInputbox extends Component {
+    render () {
+        return (
+            <TextInput
+                style = {styles.inputStyle}
+                placeholder = '해쉬태그'
+                placeholderTextColor = {'#909090'}
+            >
+            </TextInput>
+        );
+    };
 };
 
-export default InputboxID;
+export default HashtagInputbox;
