+++ src/styles/community/EmojiButton.js
@@ -1,6 +1,6 @@
 import {StyleSheet} from 'react-native';
 
-const styles = StyleSheet.create({
+const styles = StyleSheet.create ({
     buttonStyle: {
         marginLeft: 5,
         padding: 10
