+++ src/styles/post/TitleInputbox.js
@@ -1,0 +1,12 @@
+import {StyleSheet, Platform} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginTop: Platform.OS == 'ios' ? 5 : 0,
+        marginLeft: 20,
+        marginBottom: Platform.OS == "ios" ? 5 : 0,
+        padding: 10
+    }
+});
+
+export default styles;
